requirements.txt
openai==1.55.3
pydantic==2.10.2
pytest==8.3.3
python-dotenv==1.0.1
Requests==2.32.3
websockets==14.1
